+++ tofu/physics_tools/__init__.py
@@ -1,3 +1,3 @@
 
 
-from . import _runaways as runaways
+from . import runaways
